Fix window menu on macOS
Basically, this doesn't seem to be automatically handled in older versions
of macOS, which is a little annoying.

The solution, as it turns out, is quite simple after all. We just need to
request the OS for the window menu manually, which requires just a little
bit of Objective-C(++) code to get done.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>